Sales Force all variables converted to long. They were double before.

+++ src/com/marklabs/distributionCoverage/AverageMarginOffered.java
@@ -3,9 +3,9 @@ package com.marklabs.distributionCoverage;
 public class AverageMarginOffered {
 
 	private long id;
-	private double supermarket_amo;
-	private double generalStore_amo;
-	private double kiranaStore_amo;
+	private int supermarket_amo;
+	private int generalStore_amo;
+	private int kiranaStore_amo;
 
 	public long getId() {
 		return id;
@@ -15,27 +15,27 @@ public class AverageMarginOffered {
 		this.id = id;
 	}
 
-	public double getSupermarket_amo() {
+	public int getSupermarket_amo() {
 		return supermarket_amo;
 	}
 
-	public void setSupermarket_amo(double supermarketAmo) {
+	public void setSupermarket_amo(int supermarketAmo) {
 		supermarket_amo = supermarketAmo;
 	}
 
-	public double getGeneralStore_amo() {
+	public int getGeneralStore_amo() {
 		return generalStore_amo;
 	}
 
-	public void setGeneralStore_amo(double generalStoreAmo) {
+	public void setGeneralStore_amo(int generalStoreAmo) {
 		generalStore_amo = generalStoreAmo;
 	}
 
-	public double getKiranaStore_amo() {
+	public int getKiranaStore_amo() {
 		return kiranaStore_amo;
 	}
 
-	public void setKiranaStore_amo(double kiranaStoreAmo) {
+	public void setKiranaStore_amo(int kiranaStoreAmo) {
 		kiranaStore_amo = kiranaStoreAmo;
 	}
